+++ src/Stormancer.Plugins/Epic/Stormancer.Server.Plugins.Epic/EpicProfilePartBuilder.cs
@@ -150,7 +150,7 @@ namespace Stormancer.Server.Plugins.Epic
                                     data["platforms"]![EpicConstants.PLATFORM_NAME]![EpicConstants.PRODUCTUSERID] = productUserId;
                                 }
 
-                                if (user.LastPlatform == EpicConstants.PLATFORM_NAME)
+                                if (user.GetSelectedPlatformForPseudo() == EpicConstants.PLATFORM_NAME)
                                 {
                                     if (accounts.TryGetValue(accountId, out var account) && account != null && !string.IsNullOrWhiteSpace(account.DisplayName))
                                     {